move random port lookup to a separate util object
