+++ api/test/controllers/CodeSpec.scala
@@ -1,22 +1,18 @@
 package controllers
 
-import java.io.IOException
-import java.net.{InetSocketAddress, ServerSocket, Socket}
-import java.util.UUID
-import java.util.concurrent.ThreadLocalRandom
-
 import com.github.tomakehurst.wiremock.WireMockServer
-import io.apibuilder.spec.v0.models.Import
-import org.scalatestplus.play.PlaySpec
-import org.scalatestplus.play.guice.GuiceOneServerPerSuite
-import io.apibuilder.spec.v0.{models => spec}
 import com.github.tomakehurst.wiremock.client.WireMock._
 import com.github.tomakehurst.wiremock.matching.RequestPattern
-import io.apibuilder.generator.v0.models.{Invocation, InvocationForm}
 import io.apibuilder.generator.v0.models.json.{jsonReadsApibuilderGeneratorInvocationForm, jsonWritesApibuilderGeneratorInvocation}
+import io.apibuilder.generator.v0.models.{Invocation, InvocationForm}
+import io.apibuilder.spec.v0.models.Import
+import io.apibuilder.spec.v0.{models => spec}
+import org.scalatestplus.play.PlaySpec
+import org.scalatestplus.play.guice.GuiceOneServerPerSuite
 import play.api.libs.json.Json
+import util.RandomPortFinder
 
-import scala.util.{Random, Try}
+import scala.util.Try
 
 class CodeSpec extends PlaySpec with MockClient with GuiceOneServerPerSuite with db.generators.Helpers {
 
@@ -26,7 +22,7 @@ class CodeSpec extends PlaySpec with MockClient with GuiceOneServerPerSuite with
 
     "post payload containing imported services to generator" in {
 
-      val randomPort = getRandomPort
+      val randomPort = RandomPortFinder.getRandomPort
       val generatorWithService = createGenerator(createGeneratorService(createGeneratorServiceForm(s"http://localhost:$randomPort")))
 
       val generatorKey = generatorWithService.generator.key
@@ -92,29 +88,6 @@ class CodeSpec extends PlaySpec with MockClient with GuiceOneServerPerSuite with
       }
     }
 
-  }
-
-  def getRandomPort: Int = {
-    Stream
-      .continually(randomPort)
-      .dropWhile(port => !checkIfPortIsFree("localhost", port))
-      .head
-  }
-
-  private def randomPort: Int = {
-    10000 + Random.nextInt(55000)
-  }
-
-  private def checkIfPortIsFree(host: String, port: Int): Boolean = {
-    val s = new Socket()
-    try {
-      s.connect(new InetSocketAddress(host, port), 1000)
-      false
-    } catch {
-      case _: IOException => true
-    } finally {
-      s.close()
-    }
   }
 
 }
